Refactor and move button one row higher
